requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile ./requirements.in --resolver=backtracking --output-file ./requirements.txt
asgiref==3.6.0
    # via django
asttokens==2.2.1
    # via
    #   friendly-traceback
    #   stack-data
attrs==20.3.0
    # via spatula
backcall==0.2.0
    # via ipython
beautifulsoup4==4.12.2
bleach==6.0.0
    # via django-markdownify
bumpver==2023.1121
certifi==2024.2.2
    # via
    #   django-tailwind-cli
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   bumpver
    #   django-click
    #   django-typer
    #   spatula
    #   typer
colorama==0.4.6
    # via bumpver
cssselect==1.2.0
    # via spatula
dateutils==0.6.12
decorator==5.1.1
    # via ipython
distlib==0.3.8
    # via virtualenv
dj-database-url==2.0.0
    # via environs
dj-email-url==1.0.6
    # via environs
django==3.2.19
    # via
    #   dj-database-url
    #   django-ajaximage
    #   django-braces
    #   django-contrib-comments
    #   django-filter
    #   django-ical
    #   django-markdownify
    #   django-picklefield
    #   django-q2
    #   django-recurrence
    #   django-simple-open-graph
    #   django-storages
    #   django-taggit
    #   django-tailwind-cli
    #   django-thumbor
    #   django-typer
    #   djangorestframework
    #   heroicons
django-ajaximage @ git+https://github.com/trailhawks/django-ajaximage.git@ab4ceca0bed93216d317d5576be9ce41c45e657b
django-braces==1.15.0
django-cache-url==3.4.4
    # via environs
django-click==2.3.0
django-contrib-comments==2.2.0
django-favicon==0.1.4
django-filter==23.2
django-ical==1.9.1
django-markdownify==0.9.3
django-markup-deprecated==0.0.4
django-picklefield==3.1
    # via django-q2
django-q2==1.5.3
django-recurrence==1.11.1
    # via django-ical
django-robots==5.0
django-shorturls==2.0.0
django-simple-open-graph==0.3
django-storages==1.13.2
django-syncr @ git+https://github.com/trailhawks/django-syncr.git@2fac989f3538b2c28483e4a71ae713fafb5cf8b9
django-taggit==4.0.0
django-tailwind-cli==2.9.0
django-thumbor==0.5
django-typer==1.0.5
    # via django-tailwind-cli
djangorestframework==3.12.4
docopt==0.6.2
    # via num2words
environs==9.5.0
et-xmlfile==1.1.0
    # via openpyxl
executing==1.2.0
    # via
    #   friendly-traceback
    #   stack-data
feedparser==6.0.10
filelock==3.13.1
    # via virtualenv
flickrapi==2.0
friendly==0.7.21
friendly-styles==0.2
    # via friendly
friendly-traceback==0.7.61
    # via friendly
greenlet==2.0.1
    # via playwright
gunicorn==20.1.0
heroicons==2.6.0
html5lib==1.1
httplib2==0.22.0
icalendar==5.0.5
    # via django-ical
identify==2.5.35
    # via pre-commit
idna==3.4
    # via requests
ipython==7.34.0
    # via spatula
jedi==0.18.2
    # via ipython
lexid==2021.1006
    # via bumpver
libthumbor==2.0.2
    # via django-thumbor
looseversion==1.1.2
    # via bumpver
lxml==4.9.2
    # via spatula
markdown==3.4.3
    # via django-markdownify
markdown-it-py==2.2.0
    # via rich
marshmallow==3.19.0
    # via environs
matplotlib-inline==0.1.6
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
micawber==0.5.4
nodeenv==1.8.0
    # via pre-commit
num2words==0.5.12
numpy==1.24.3
    # via pandas
oauthlib==3.2.2
    # via requests-oauthlib
openpyxl==3.1.2
    # via spatula
packaging==23.1
    # via marshmallow
pandas==2.0.1
parso==0.8.3
    # via jedi
pathlib2==2.3.7.post1
    # via bumpver
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.5.0
    # via django-ajaximage
platformdirs==3.5.0
    # via
    #   friendly
    #   virtualenv
playwright==1.33.0
pre-commit==3.6.2
prompt-toolkit==3.0.38
    # via ipython
psycopg2-binary==2.9.6
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via
    #   friendly-traceback
    #   stack-data
pyee==9.0.4
    # via playwright
pygments==2.15.1
    # via
    #   friendly
    #   friendly-styles
    #   ipython
    #   rich
pyparsing==3.0.9
    # via httplib2
python-dateutil==2.8.2
    # via
    #   dateutils
    #   django-recurrence
    #   icalendar
    #   pandas
python-dotenv==1.0.0
    # via environs
python-slugify==8.0.1
pytz==2023.3
    # via
    #   dateutils
    #   django
    #   icalendar
    #   pandas
pyyaml==6.0.1
    # via pre-commit
requests==2.30.0
    # via
    #   flickrapi
    #   requests-oauthlib
    #   requests-toolbelt
    #   scrapelib
requests-oauthlib==1.3.1
    # via flickrapi
requests-toolbelt==1.0.0
    # via flickrapi
rich==13.3.5
    # via friendly
scrapelib==2.1.0
    # via spatula
setuptools==69.2.0
    # via
    #   django-shorturls
    #   gunicorn
    #   ipython
    #   nodeenv
sgmllib3k==1.0.0
    # via feedparser
shellingham==1.5.4
    # via django-typer
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   django-shorturls
    #   flickrapi
    #   friendly-traceback
    #   html5lib
    #   libthumbor
    #   pathlib2
    #   python-dateutil
soupsieve==2.4.1
    # via beautifulsoup4
spatula==0.9.0
sqlparse==0.4.4
    # via django
stack-data==0.6.2
    # via friendly-traceback
text-unidecode==1.3
    # via python-slugify
tinycss2==1.1.1
    # via bleach
titlecase==2.4
toml==0.10.2
    # via bumpver
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
typer==0.9.0
    # via django-typer
typing-extensions==4.5.0
    # via
    #   dj-database-url
    #   pyee
    #   typer
tzdata==2023.3
    # via pandas
urllib3==2.0.2
    # via requests
uwsgi==2.0.21
virtualenv==20.23.0
    # via pre-commit
wcwidth==0.2.6
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   html5lib
    #   tinycss2
whitenoise==6.4.0
